Factor non-Abelian GI operator tables per vertex
Replace the dense plaquette/hopping matrix tables (row index dense over
block-id tuples, O(n_blocks^4) per plaquette) with per-vertex Schur factor
tables: every string/loop operator matrix element is an exact product of
corner-local scalars, stored as one VertexFactorTable per (site, role,
orientation) keyed by (input block, new touched-leg sectors). Daggered links
use sector-swapped conjugated couplings of the same operator irrep, so the
construction holds in any basis of any compact group; the dual irrep never
enters operator tensors.

- factors.py: generic oracles, cached NumPy factor extraction, table
  builders with build-time ring == prod(lambda) split asserts, and the
  shared window conventions (key legs / fusion orientation per moved leg).
- kernels.py: multilinear fold (one lambda-weighted gather-sum per corner,
  one window contraction per valid links combo), ancestral proposal with
  exact Hastings ratio Z(in)/Z(out), vectorized window-combo enumeration
  (3x3 j_max=1 transition trace: 62k -> 11.5k jaxpr eqns, 24s -> 0.8s).
- initial.py: DP-weighted initial sampling (backward completion counts +
  forward walk) with conserved-particle-number matter partitions.
- contraction.py: single-source sample codec and sampled-block row MPOs.
- su2.py/su3.py shrink to group primitives (irreps, couplings, intertwiner
  enumeration); all operator-table logic is group-generic.
- Validation gate tests/test_non_abelian_factors.py anchors to external
  truths (Haar plaquette saturation, exact Gauss-sector meson hop, Gauss
  invariance of matter vertices); shared exact-enumeration helpers in
  tests/nonabelian_exact.py; ED baselines re-anchored to the canonical
  matrix elements (su2 3x3: -4.8234564111911995; su3 2x2:
  -0.8509840232358125).
- Specs: factor-table design and the graded sampled-sector PEPS design
  (unifying fermionic statistics and gauge structure as sector metadata).

Full suite: 256 passed + 76 subtests.

+++ src/vmc/peps/standard/kernels.py
@@ -83,7 +83,7 @@ def _assemble_log_derivatives(
         for col in range(n_cols)
     ]
     active_slice_indices = jnp.repeat(
-        config_state.astype(jnp.int8),
+        config_state.astype(jnp.int16),
         params_per_site,
         axis=0,
         total_repeat_length=total_active_params,
